add tests for book and categories

+++ src/main/java/mate/academy/demo/service/BookServiceImpl.java
@@ -46,6 +46,9 @@ public class BookServiceImpl implements BookService {
 
     @Override
     public void deleteById(Long id) {
+        if (!bookRepository.existsById(id)) {
+            throw new EntityNotFoundException("Can't find book with id: " + id);
+        }
         bookRepository.deleteById(id);
     }
